+++ com.unity.render-pipelines.core/Runtime/Common/CoreAttributes.cs
@@ -1,5 +1,9 @@
 using System;
 
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
+
 namespace UnityEngine.Rendering
 {
     /// <summary>