+++ com.unity.render-pipelines.core/Runtime/Utilities/CoreUtils.cs
@@ -1,8 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
+using System.Reflection;
 using UnityEngine.Experimental.Rendering;
 using UnityEngine.Rendering.RendererUtils;
+
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 
 namespace UnityEngine.Rendering
 {